Changes for int to be set as Double

+++ eResourceWeb/Models/BaseModel.cs
@@ -8,7 +8,7 @@ namespace eResourceWeb.Models
     public abstract class BaseModel
     {
         public DateTime Last_Updated { get; set; }
-        public int Is_Deleted { get; set; }
+        public Double Is_Deleted { get; set; }
         public string ModifiedBy { get; set; }
     }
 }
